clarify intentional conversion of long to signed int

+++ xmchord/helper/textual.cc
@@ -39,9 +39,7 @@ int Textual::StrPos(char *hay, char *needle, int offset) {
 
   char *p = strstr(haystack, needle);
 
-  return (p)
-     ? p - haystack + offset
-     : -1;
+  return p ? (int)(p - haystack + offset) : -1;
 }
 
 bool Textual::Contains(const std::string &haystack, const char *needle) {
